[C++] Homework 1
Add Wall ability -> The player is able to shoot a cubic projectile
which upon impact with something other than the player that shot it
should create a wall at that position, the wall should be facing towards
the player, as that seems more intuitive to me, the player can also
reactivate the ability in order to create a wall at the current
position of the projectile. Add mesh projectile
since the only projectile we have in the project is hardcoded to be
spherical, MeshProjectile.h has the advantage that it can be bound to some
sort of static mesh from blueprints. Remove some of the repetitive code from
TopDownARPG/TopDownARPGPlayerController.cpp as it annoyed me.
Add more inputs, in order to use more abilities.